backend/requirements.txt
fastapi==0.100.1
uvicorn==0.23.1
pytest==7.4.0
httpx==0.24.1
pymongo==4.4.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
passlib==1.7.4
python-jose==3.3.0
